refactor : 즐겨찾기 페이지 폴더명 변경
