+++ src/pages/sharedPage/SharedPage.jsx
