+++ src/pages/sharedPage/SharedPage.module.css
